# routes/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Music Stram descritption",
        "version": "1.0.0",
        "title": "Music Stream",
        "termsOfService": "",
        "contact": {
            "email": "hspharwinder@gmail.com"
        }
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {  
...
        "/forgetPassword": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Recover user password by email",
                "description": "",
                "operationId": "forgetPassword",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "email",
                        "description": "",
                        "required": true,
                        "type": "string"                        
                    }
                ],
                "responses": {
                    "200": {
                        "description": "responsed json array of user",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Email not registered"
                    }
                }
            }
        },
...
        "/approveToArtist": {
            "post": {
                "tags": [
                    "Approval"
                ],
                "summary": "Give approval to artist, by admin",
                "description": "Give approval to artist by updating username and passord to the artist, process will be done by admin only",
                "operationId": "approveToArtist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "artistId",
                        "description": "Artist id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "userName",
                        "description": "User name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "password",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "responsed json Array of user",
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Fail to update record"
                    }
                }
            }
        },
...
        "/checkwishlist": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Check wether record is present in wishlist or not, pass user id, media id",
                "description": "",
                "operationId": "checkwishlist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "userId",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "mediaId",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record present"
                    },
                    "400": {
                        "description": "Record not present"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {  
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "type": {
                    "type": "integer",
                    "format": "int(11)"
                },
                "image": {
                    "type": "string",
                    "description": "First upload your image through 'filePost' endpoints and paste here responed image path"                   
                },
                "status": {
                    "type": "integer",
                    "format": "tinyInt"
                    
                },
                "phone_no": {
                    "type": "string",
                    "format": "mobile"                    
                },
                "description": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "User"
            }
        },
...
        "Artist": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "type": {
                    "type": "integer",
                    "format": "int(11)"
                },
                "image": {
                    "type": "string",
                    "description": "First upload your image through 'filePost' endpoints and paste here responed image path"
                },
                "status": {
                    "type": "integer",
                    "format": "tinyInt"
                },
                "phone_no": {
                    "type": "string",
                    "format": "mobile"
                },
                "description": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Artist"
            }
        },
...
    },
...
}